requirements.txt
astroid==2.4.1
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.3
defusedxml==0.6.0
Django==2.2
django-allauth==0.41.0
django-filter==2.2.0
django-rest-auth==0.9.5
django-webpack-loader==0.6.0
djangorestframework==3.11.0
flake8==3.8.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.0.0
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.5.2
python-dateutil==2.8.1
python-http-client==3.2.4
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
sendgrid==6.1.1
six==1.14.0
sqlparse==0.3.0
toml==0.10.1
urllib3==1.25.8
wrapt==1.12.1
